separate LoginCode components and update imports
