+++ src/components/Login/RequestLoginCode/index.js
@@ -1,0 +1,61 @@
+import React, { useState, useContext } from 'react';
+
+import AuthContext from '../../../context/Authentication';
+import { useSignOut } from '../../../hooks/Authentication';
+import { FinallyMessage } from '../../Forms/FinallyMessage';
+import { EnterLoginCode } from '../EnterLoginCode';
+import {
+  CTAButtons,
+  CTAButtonContainer,
+  CTAButton,
+} from '../../Layout/CTAButton';
+
+export const RequestLoginCode = ({ children, buttonText }) => {
+  const { customUserData: userData } = useContext(AuthContext);
+  const [confirmSendCode, setConfirmSendCode] = useState(false);
+  const signOut = useSignOut();
+
+  if (!confirmSendCode) {
+    // TODO only show email as fallback
+    return (
+      <FinallyMessage type="success">
+        {children ? (
+          children
+        ) : (
+          <p>
+            Du bist schon angemeldet
+            {userData.username
+              ? ` als ${userData.username}`
+              : userData.email
+              ? `mit dem email ${userData.email}`
+              : ''}
+            .
+          </p>
+        )}
+        <CTAButtons>
+          <CTAButtonContainer>
+            <CTAButton onClick={() => setConfirmSendCode(true)} type="button">
+              {buttonText || 'Einloggen'}
+            </CTAButton>
+          </CTAButtonContainer>
+          <CTAButtonContainer>
+            <CTAButton onClick={signOut} type="button">
+              Abmelden
+            </CTAButton>
+          </CTAButtonContainer>
+        </CTAButtons>
+      </FinallyMessage>
+    );
+  }
+
+  // If there is a temporary email, show EnterLoginCode
+  return (
+    <EnterLoginCode>
+      <p>
+        {' '}
+        Um dich zu identifizieren, haben wir dir einen Code per E-Mail
+        geschickt. Bitte gib diesen ein:
+      </p>
+    </EnterLoginCode>
+  );
+};
