+++ src/hooks/Authentication/index.js
@@ -113,6 +113,10 @@ const signOut = async ({ setCognitoUser, setUserId, setIsAuthenticated }) => {
 
     await Auth.signOut();
 
+    // Get scroll location
+    const scrollLocation = window.scrollY;
+    console.log(scrollLocation);
+
     // Remove URL params if existing
     const params = querystring.parse(window.location.search);
     if (params.userId) {
@@ -122,6 +126,9 @@ const signOut = async ({ setCognitoUser, setUserId, setIsAuthenticated }) => {
       window.location.pathname
     }${querystring.stringify(params)}`;
     navigate(newUrl, { replace: true });
+
+    // Set scroll location
+    window.scroll(0, scrollLocation);
 
     // Update user state
     setCognitoUser(null);
